requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Libraries
openai==1.3.0
langchain==0.0.346
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0

# Vector Database
chromadb==0.4.13
# Alternative options:
# weaviate-client==3.24.0
# pinecone-client==2.2.1

# Audio Processing
librosa==0.10.0
soundfile==0.12.1
moviepy==1.0.3
pydub==0.25.1

# Speech Recognition
openai-whisper==20231117
# Alternative:
# google-cloud-speech==2.21.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
# Optional databases:
# psycopg2-binary==2.9.9  # PostgreSQL
# pymongo==4.6.0  # MongoDB

# Caching & Storage
redis==5.0.1
# Optional:
# boto3==1.34.0  # AWS S3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1
httpx==0.25.2

# Logging & Monitoring
python-json-logger==2.0.7
# Optional:
# sentry-sdk==1.38.0

# Performance
numpy==1.24.3
scipy==1.11.4
pandas==2.1.3
